[Bugfix]: Don't output named exports that are javascript reserved words.
* Filter out classnames that are reserved javascript keywords in either
  es5 or es6 javascript.
* Output a warning for each classname filtered out, with a suggestion to
  use default exports instead if the filtered out classname was required
  as part of the export.

Use case: using bootstrap-sass (v3), the classname "in" is created.
Since this is a reserved keyword in javascript, the named export
definition file that was being created was invalid, breaking the ability
to compile the project. However, the user is not likely to be using this
as 'styles.in' anywhere, so removing the invalid export does not
prevent the user from access to the classnames that they do use. The
warning message both informs the user that a classname was excluded,
and provides instructions on how to work around the issue if they did in
fact need to access the classname as 'styles.in' in their code.