[Feature]: Implement ability to define alias prefixes.
Implements #9.

When using Webpack with sass-loader to compile scss files, the standard
import syntax is `~package-name/path/to/file`, where `~` is then
replaced with the string `node_modules/`. In this case, it can become
very bloated to have to individually specify each imported scss file
alias for these imports. Being able to define a prefix value to replace
within the filename string means that a single prefix can be used to
match the sass-loader/webpack import syntax for all scss imports.