feat(option): allow to ignore initial build in watch
This adds an option to ignore the initial run when using the watch param. This is useful when running concurrently with some other script (another watch, for instance), but this should finish first the first time. So you run this first, then run watch with ignore initial true. This is named `ignoreInitial` in consistence with `chokidar`.