feat: add ability to delete old .d.ts files for *.scss files
This commit adds the ability to remove old .d.ts files while
running the cli with the watch flag (`--watch`). The script leverages
the unlink fs event to know when a file has been deleted and then
remove the analagous .d.ts file. The commit tries to make sure all
of the removal logic is synchronized.