chore: remove extra check if .scss file does exist
Remove an extra check within the removeScssTypeFile to avoid
removing a type file if the .scss file stil exists. This function
will assume the .scss file is already remove before proceeding to
remove the type file definition. In addition, this commits
removes the unit test that tested the functionality of not removing
existing .scss files.