fix(cli): sass error handling
This stops ignoring errors generated by the sass compiler (as well as
any other potential errors generated during the process of writing to
disk). Instead, errors are now always thrown to the top of the stack,
they are logged, and properly set the exit code. Any errors that include
file info are assumed to be sass errors, and some additional formatting
applied to improve debuggability.